switch to f64

+++ src/camera.rs
@@ -17,7 +17,7 @@ impl Camera {
             origin: Vec3::new(0.0, 0.0, 0.0),
         }
     }
-    pub fn get_ray(&self, u: f32, v: f32) -> Ray {
+    pub fn get_ray(&self, u: f64, v: f64) -> Ray {
         Ray::new(
             &self.origin,
             &(&(&self.lower_left_corner + &(&(&self.horizontal * u) + &(&self.vertical * v)))
